requirements.txt
python-telegram-bot==11.1.0
redis==3.2.1